add workspace tests ci/cd + change stwo main repo
