+++ tests/src/lib.rs
@@ -1,0 +1,1 @@
+pub mod e2e_test;
